third_party/blink/web_tests/platform/generic/fast/css/font-face-unicode-range-overlap-load-expected.txt

Tests that only necessary fonts are loaded when font faces have overlapping unicode ranges.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS faces.length is 6
PASS faces[0].status is "unloaded"
PASS faces[1].status is "loaded"
PASS faces[2].status is "loaded"
PASS faces[3].status is "unloaded"
PASS faces[4].status is "unloaded"
PASS faces[5].status is "loaded"
PASS successfullyParsed is true

TEST COMPLETE


I

J

K
